requirements.txt
diff-match-patch==20121119
dj-database-url==0.4.0
dj-static==0.0.6
Django==1.9.2
django-easy-pdf==0.1.0
django-import-export==0.4.2
django-solo==1.1.2
django-toolbelt==0.0.1
gunicorn==19.4.5
html5lib==0.9999999
Pillow==3.1.1
psycopg2==2.6.1
PyPDF2==1.25.1
python-dateutil==2.4.2
reportlab==2.7
six==1.10.0
static3==0.6.1
tablib==0.11.2
wheel==0.24.0
xhtml2pdf==0.0.6
